+++ src/renderer/VideoJS.tsx
@@ -1,12 +1,25 @@
-import { RendererVideo, TAppState, VideoPlayerSettings } from 'main/types';
+import {
+  DeathMarkers,
+  RendererVideo,
+  TAppState,
+  VideoPlayerSettings,
+} from 'main/types';
 import React from 'react';
 import videojs from 'video.js';
 import 'videojs-hotkeys';
 import 'video.js/dist/video-js.css';
-import 'videojs-markers/dist/videojs-markers';
-import 'videojs-markers/dist/videojs.markers.css';
-import Player from 'video.js/dist/types/player';
-import { addVideoMarkers } from './rendererutils';
+import { ConfigurationSchema } from 'main/configSchema';
+import { VideoCategory } from 'types/VideoCategory';
+import {
+  addMarkerDiv,
+  convertNumToDeathMarkers,
+  getAllDeathMarkers,
+  getEncounterMarkers,
+  getMarkerDiv,
+  getOwnDeathMarkers,
+  getRoundMarkers,
+  removeMarkerDiv,
+} from './rendererutils';
 
 interface IProps {
   // eslint-disable-next-line react/no-unused-prop-types
@@ -15,18 +28,82 @@ interface IProps {
   key: string;
   video: RendererVideo;
   setAppState: React.Dispatch<React.SetStateAction<TAppState>>;
+  config: ConfigurationSchema;
 }
 
 const ipc = window.electron.ipcRenderer;
 
 export const VideoJS = (props: IProps) => {
+  const { video, setAppState, config } = props;
   const videoRef: any = React.useRef(null);
   const playerRef: any = React.useRef(null);
-  const { video, setAppState } = props;
+  const markerDivs: React.MutableRefObject<HTMLDivElement[]> = React.useRef([]);
+
+  // This is a bit sneaky, but we create a dummy state variable to give us a
+  // mechanism to force react to re-render on demand, which we use to apply
+  // marker config changes.
+  const [, forceReRender] = React.useState<number>(0);
+
+  /**
+   * Remove any existing timeline markers, then collate a list of all the
+   * markers based on the category and current config.
+   *
+   * Finally, add the markers to the timelime. This code is not very reacty, I
+   * blame VideoJS.
+   *
+   * We pass cfg in here as an argument to prevent remembering old state. I
+   * don't really understand why that is required.
+   */
+  const processMarkers = () => {
+    markerDivs.current.forEach(removeMarkerDiv);
+    markerDivs.current = [];
+
+    const progressBar = document.querySelector('.vjs-progress-holder');
+
+    if (
+      !playerRef.current ||
+      !progressBar ||
+      Number.isNaN(playerRef.current.duration())
+    ) {
+      return;
+    }
+
+    const duration = playerRef.current.duration();
+    const { width } = progressBar.getBoundingClientRect();
+    const deathMarkerConfig = convertNumToDeathMarkers(config.deathMarkers);
+
+    if (deathMarkerConfig === DeathMarkers.ALL) {
+      getAllDeathMarkers(video)
+        .map((m) => getMarkerDiv(m, duration, width))
+        .forEach((m) => markerDivs.current.push(m));
+    } else if (deathMarkerConfig === DeathMarkers.OWN) {
+      getOwnDeathMarkers(video)
+        .map((m) => getMarkerDiv(m, duration, width))
+        .forEach((m) => markerDivs.current.push(m));
+    }
+
+    const isMythicPlus = video.category === VideoCategory.MythicPlus;
+
+    if (isMythicPlus && config.encounterMarkers) {
+      getEncounterMarkers(video)
+        .map((m) => getMarkerDiv(m, duration, width))
+        .forEach((m) => markerDivs.current.push(m));
+    }
+
+    const isSoloShuffle = video.category === VideoCategory.SoloShuffle;
+
+    if (isSoloShuffle && config.roundMarkers) {
+      getRoundMarkers(video)
+        .map((m) => getMarkerDiv(m, duration, width))
+        .forEach((m) => markerDivs.current.push(m));
+    }
+
+    markerDivs.current.forEach((m) => addMarkerDiv(m));
+  };
 
   /**
    * Get video player settings initially when the component is loaded. We store
-   * as a variable in main rather than in config It's fine if this is lost when
+   * as a variable in main rather than in config. It's fine if this is lost when
    * the app is restarted.
    */
   const videoPlayerSettings = ipc.sendSync('videoPlayerSettings', [
@@ -40,13 +117,6 @@ export const VideoJS = (props: IProps) => {
   const handleVideoPlayerVolumeChange = (volume: number, muted: boolean) => {
     const soundSettings: VideoPlayerSettings = { volume, muted };
     ipc.sendMessage('videoPlayerSettings', ['set', soundSettings]);
-  };
-
-  /**
-   * Add the markers when the video player is ready.
-   */
-  const onVideoPlayerReady = (player: Player) => {
-    addVideoMarkers(video, player);
   };
 
   React.useEffect(() => {
@@ -90,7 +160,6 @@ export const VideoJS = (props: IProps) => {
       playerRef.current.on('ready', () => {
         playerRef.current.muted(videoPlayerSettings.muted);
         playerRef.current.volume(videoPlayerSettings.volume);
-        onVideoPlayerReady(playerRef.current);
       });
 
       playerRef.current.on('volumechange', () => {
@@ -98,6 +167,14 @@ export const VideoJS = (props: IProps) => {
           playerRef.current.volume(),
           playerRef.current.muted()
         );
+      });
+
+      playerRef.current.on('play', () => {
+        forceReRender((n) => n + 1);
+      });
+
+      playerRef.current.on('playerresize', () => {
+        forceReRender((n) => n + 1);
       });
 
       playerRef.current.on('fullscreenchange', () => {
@@ -120,6 +197,8 @@ export const VideoJS = (props: IProps) => {
       }
     };
   }, [playerRef]);
+
+  processMarkers();
 
   return (
     <div data-vjs-player className="video-container">